requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.104.1
greenlet==3.0.1
h11==0.14.0
idna==3.6
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
